+++ include/visualise.h
@@ -6,13 +6,13 @@
     Return à normalize value based on
     a coordinate and his max and min value.
 */
-double normalize_to_screen(double coord, double coord_max, double coord_min);
+double normalize_to_screen(float coord, float coord_max, float coord_min);
 
 /* 
     Based on a node_list, get x=lattitude y=longitude min max value.
     return 0 if no problems, otherwise 1.
 */
-int get_xy_min_max(list_t* node_list, double* x_max, double* x_min, double* y_max, double* y_min);
+int get_xy_min_max(list_t* node_list, float* x_max, float* x_min, float* y_max, float* y_min);
 
 
 /*
